Fix ISO14443-4 RX truncation of frames >=32 bytes (unwrap_frame uint8_t return)
nfc_tag_14a_unwrap_frame() computes the de-parity'd data-bit count in a size_t
but returns uint8_t. For any received frame of 32 bytes or more the data-bit
count exceeds 255, so the uint8_t return wraps it modulo 256, silently
truncating the frame the caller (nfc_tag_14a_data_process) then processes.

Concretely a 44-byte reader->tag command is 352 data bits; 352 & 0xFF = 96 bits
= 12 bytes, so only the first 12 bytes reach the tag/APDU layer. This breaks any
ISO14443-4 (T=CL) tag emulation that receives a command of 32+ bytes -- e.g. an
EMV GET PROCESSING OPTIONS carrying a full PDOL, where the card then answers
6700 (wrong length).

Verified on nRF52840 hardware: the NFCT RXD.AMOUNT register reports the correct
length at every frame size (396 for a 44-byte frame); only the truncated return
value was wrong. Changing the return type to uint16_t makes large commands
arrive intact.

Reproduced deterministically by having one Chameleon (RC522 reader) send
commands of increasing size to another (NFCT emulator): received length was
(frame length mod 32) before the fix, full length after.

+++ firmware/application/src/rfid/nfctag/hf/nfc_14a.h
@@ -124,7 +124,7 @@ bool nfc_tag_14a_checks_crc(uint8_t *pbtData, size_t szLen);
 
 // 14A frame combination
 uint8_t nfc_tag_14a_wrap_frame(const uint8_t *pbtTx, const size_t szTxBits, const uint8_t *pbtTxPar, uint8_t *pbtFrame);
-uint8_t nfc_tag_14a_unwrap_frame(const uint8_t *pbtFrame, const size_t szFrameBits, uint8_t *pbtRx, uint8_t *pbtRxPar);
+uint16_t nfc_tag_14a_unwrap_frame(const uint8_t *pbtFrame, const size_t szFrameBits, uint8_t *pbtRx, uint8_t *pbtRxPar);
 
 // 14A communication control
 void nfc_tag_14a_sense_switch(bool enable);
